LLM/financial_analyzer/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
langchain>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
requests>=2.31.0
PyMuPDF>=1.23.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
